Add files via upload
Improvements to the backup file - now backs-up on "sc -fe"

Improvements to "sc -fs" for future, and long-term use.  If DB is bigger than a certain number of entries, "less" will be used instead of "cat" to display the DB.  After a certain size, the DB will be very annoying.